Bloc3 - Fraud detection by ML/Projet/Architecture/localModel/requirements copy.txt
pandas==2.3.3
numpy==2.2.6
mlflow==2.21.3
scikit-learn==1.4.2
python-dotenv==1.2.1
boto3==1.42.3
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
